Suppress warnings and fix other minor issues
